Refactored CanvasNode class to use xaml Removed unneeded xaml for Node Removed component panel Added node button on toolbar
