+++ NetworkObservability/CanvasNode.xaml
@@ -1,0 +1,49 @@
+﻿<UserControl x:Class="NetworkObservability.CanvasNode"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             mc:Ignorable="d" 
+             DataContext="{Binding RelativeSource={RelativeSource Self}}"
+             d:DesignHeight="100" d:DesignWidth="100">
+    <StackPanel Orientation="Vertical" ToolTip="A node that can be dragged onto the canvas to create a graph" VerticalAlignment="Center">
+        <Button Margin="0 0 0 0" VerticalAlignment="Center">
+            <Button.Template>
+                <ControlTemplate>
+                    <Rectangle Height="20" Width="20" Margin="0" RenderTransformOrigin="0.5,0.5" HorizontalAlignment="Center">
+                        <Rectangle.RenderTransform>
+                            <TransformGroup>
+                                <RotateTransform Angle="45"/>
+                            </TransformGroup>
+                        </Rectangle.RenderTransform>
+                        <Rectangle.Style>
+                            <Style TargetType="{x:Type Rectangle}">
+                                <Setter Property="Fill" Value="Black" />
+                                <Style.Triggers>
+                                    <Trigger Property="IsMouseOver" Value="True">
+                                        <Setter Property="Fill" Value="#444444" />
+                                    </Trigger>
+                                    <DataTrigger Binding="{Binding IsObserver}" Value="True">
+                                        <Setter Property="Fill" Value="#5fba7d"/>
+                                    </DataTrigger>
+                                </Style.Triggers>
+                            </Style>
+                        </Rectangle.Style>
+                    </Rectangle>
+                </ControlTemplate>
+            </Button.Template>
+        </Button>
+        <TextBlock HorizontalAlignment="Center" FontWeight="Bold" FontSize="11" VerticalAlignment="Center" Margin="0 5 0 0" Text="{Binding Label}">
+            <TextBlock.Style>
+                <Style TargetType="TextBlock">
+                    <Setter Property="Visibility" Value="Visible"/>
+                    <Style.Triggers>
+                        <DataTrigger Binding="{Binding Label}" Value="N0000000">
+                            <Setter Property="Visibility" Value="Collapsed"/>
+                        </DataTrigger>
+                    </Style.Triggers>
+                </Style>
+            </TextBlock.Style>
+        </TextBlock>
+    </StackPanel>
+</UserControl>
