+++ NetworkObservability/MainWindow.xaml
@@ -4,7 +4,6 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:custom="clr-namespace:NetworkObservability"
-        xmlns:resources="clr-namespace:NetworkObservability.resources"
         mc:Ignorable="d"
         WindowState="Maximized"
         Title="MainWindow"  AllowDrop="True">
@@ -34,24 +33,26 @@
                     <ColumnDefinition Width="2*" />
                 </Grid.ColumnDefinitions>
                 <!-- Left toolbar -->
-                <ToolBar Grid.Column="0" Background="{DynamicResource {x:Static SystemColors.ControlBrushKey}}">
-                    <StackPanel Orientation="Horizontal">
-                        <Button Margin="5,0,0,4" Content="Start" Width="70" HorizontalAlignment="Left" Style="{DynamicResource ButtonStyleForStart}" Click="Start_Click" VerticalAlignment="Bottom"/>
-                        <Button x:Name="___Button___Pause_" Margin="5 0 0 0" Content="Start" Width="25" HorizontalAlignment="Left" Style="{DynamicResource ButtonStylePause}" IsEnabled="False" />
-                        <Button x:Name="___Button___Stop_" Margin="5 0 0 0" Content="Start" Width="25" HorizontalAlignment="Left" Style="{DynamicResource ButtonStyleStop}" IsEnabled="False" />
+                <ToolBar Grid.Column="0" Background="{DynamicResource {x:Static SystemColors.ControlBrushKey}}" ToolBarTray.IsLocked="True" Height="50">
+                    <StackPanel Orientation="Horizontal" VerticalAlignment="Center" HorizontalAlignment="Center" Margin="15 0 0 0">
+                        <Button Content="Start" Width="70" HorizontalAlignment="Left" Style="{DynamicResource ButtonStyleForStart}" Click="Start_Click"/>
+                        <Button x:Name="___Button___Pause_" Margin="5 0 0 0" Content="Start" Width="25" HorizontalAlignment="Left" Style="{DynamicResource ButtonStylePause}" IsEnabled="False"/>
+                        <Button x:Name="___Button___Stop_" Margin="5 0 0 0" Content="Start" Width="25" HorizontalAlignment="Left" Style="{DynamicResource ButtonStyleStop}" IsEnabled="False"/>
                     </StackPanel>
                 </ToolBar>
                 <!-- Left toolbar -->
                 <!-- Right toolbar -->
-                <ToolBar Grid.Column="2" Background="{DynamicResource {x:Static SystemColors.ControlBrushKey}}">
-                    <StackPanel Orientation="Horizontal" Margin="5 0 0 0">
-                        <Label Margin="0 0 0 0" Content="Arc type:" />
-                        <ComboBox x:Name="ArcType" Margin="3 0 0 0" Width="150" Background="{DynamicResource {x:Static SystemColors.ControlLightLightBrushKey}}" BorderBrush="#00ACACAC" >
+                <ToolBar Grid.Column="2" Background="{DynamicResource {x:Static SystemColors.ControlBrushKey}}" Height="50" ToolBarTray.IsLocked="True" >
+                    <StackPanel Orientation="Horizontal" VerticalAlignment="Center" Margin="35 0 0 0">
+                        <Label Content="Node:" VerticalContentAlignment="Center" FontWeight="Bold"></Label>
+                        <custom:CanvasNode x:Name="canvasNodeButton" Margin="20 0 20 0" VerticalContentAlignment="Center"></custom:CanvasNode>
+                        <Label Content="Arc type:" VerticalContentAlignment="Center" FontWeight="Bold"/>
+                        <ComboBox x:Name="ArcType" Height="25" Width="150" Background="{DynamicResource {x:Static SystemColors.ControlLightLightBrushKey}}" BorderBrush="#00ACACAC" >
                             <ComboBoxItem x:Name="UndirectedArc" Content="Undirected lines" Selector.IsSelected="True" Background="{DynamicResource ToolBarButtonPressed}" />
                             <ComboBoxItem Content="Directed lines"/>
                         </ComboBox>
                         <!-- Add the image label for the arrows here! -->
-                        <Label Margin="2 0 0 0" Width="150" />
+                        
                         <Button x:Name="___Button___Delete_" Margin="10 0 0 0" Content="Start" Width="70" HorizontalAlignment="Left" Style="{DynamicResource ButtonStyleDelete}" />
                     </StackPanel>
                 </ToolBar>
@@ -116,19 +117,6 @@
                 <GridSplitter VerticalAlignment="Center" Grid.Row="1" Height="5" Background="#FFBCBCBC" HorizontalAlignment="Stretch"/>
 
                 <TabControl Grid.Row="2" TabStripPlacement="Bottom" BorderThickness="1,0,1,1" >
-                    <TabItem Header="Components" >
-                        <ScrollViewer VerticalScrollBarVisibility="Auto" Margin="0" Padding="0" Background="LightBlue" >
-                            <UniformGrid Columns="6" VerticalAlignment="Center">
-                                <resources:VisibleObserver x:Name="visibleObserver"></resources:VisibleObserver>
-                                <resources:InvisibleObserver x:Name="invisibleObserver"></resources:InvisibleObserver>
-                                <resources:VisibleNode x:Name="visibleNode"></resources:VisibleNode>
-                                <resources:InvisibleNode x:Name="invisibleNode"></resources:InvisibleNode>
-                                <resources:EndNode x:Name="endNode"></resources:EndNode>
-                                <resources:CustomNode x:Name="customNode"></resources:CustomNode>
-                            </UniformGrid>
-                        </ScrollViewer>
-                    </TabItem>
-
                     <TabItem Header="Log" Name="logTab">
                         <ScrollViewer VerticalScrollBarVisibility="Auto">
                             <Grid Height="Auto">
