+++ NetworkObservability/NetworkObservability.csproj
@@ -61,25 +61,10 @@
       <SubType>Designer</SubType>
     </ApplicationDefinition>
     <Compile Include="CanvasGraph.cs" />
-    <Compile Include="resources\CustomNode.xaml.cs">
-      <DependentUpon>CustomNode.xaml</DependentUpon>
+    <Compile Include="CanvasNode.xaml.cs">
+      <DependentUpon>CanvasNode.xaml</DependentUpon>
     </Compile>
-    <Compile Include="resources\EndNode.xaml.cs">
-      <DependentUpon>EndNode.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="resources\InvisibleNode.xaml.cs">
-      <DependentUpon>InvisibleNode.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="resources\InvisibleObserver.xaml.cs">
-      <DependentUpon>InvisibleObserver.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="resources\VisibleNode.xaml.cs">
-      <DependentUpon>VisibleNode.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="resources\VisibleObserver.xaml.cs">
-      <DependentUpon>VisibleObserver.xaml</DependentUpon>
-    </Compile>
-    <Page Include="resources\CustomNode.xaml">
+    <Page Include="CanvasNode.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -100,18 +85,6 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Page Include="resources\EndNode.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="resources\InvisibleNode.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="resources\InvisibleObserver.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
     <Page Include="resources\Toolbar\PauseButton.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
@@ -124,17 +97,8 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="resources\VisibleNode.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="resources\VisibleObserver.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CanvasNode.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
